bug fix with empty numpy array in headerlet summary
added initial unittest for headerlet module.

print_summary, called by headerlet_summary, was erroring
when numpy now raises an exception for calling max() on an
empty array.  Fix was to default to 0 if the length of the
array was to be 0.

further expand documentation for headerlets

pep changes for inits